+++ src/Entities/Traits/ClientTrait.php
@@ -22,6 +22,11 @@ trait ClientTrait
     protected $redirectUri;
 
     /**
+     * @var bool
+     */
+    protected $isConfidential = false;
+
+    /**
      * Get the client's name.
      *
      * @return string
@@ -42,5 +47,15 @@ trait ClientTrait
     public function getRedirectUri()
     {
         return $this->redirectUri;
+    }
+
+    /**
+     * Returns true if the client is confidential.
+     *
+     * @return bool
+     */
+    public function isConfidential()
+    {
+        return $this->isConfidential;
     }
 }